feat: prepare android YamapPolygonManager for new arch

+++ src/spec/PolygonNativeComponent.ts
@@ -1,5 +1,5 @@
 import codegenNativeComponent from 'react-native/Libraries/Utilities/codegenNativeComponent';
-import {BubblingEventHandler, Double, Int32} from 'react-native/Libraries/Types/CodegenTypes';
+import {BubblingEventHandler, Double, Float, Int32} from 'react-native/Libraries/Types/CodegenTypes';
 import {ViewProps} from 'react-native';
 
 interface Point {
@@ -10,7 +10,7 @@ interface Point {
 export interface PolygonNativeProps extends ViewProps {
   fillColor?: Int32;
   strokeColor?: Int32;
-  strokeWidth?: Double;
+  strokeWidth?: Float;
   onPress?: BubblingEventHandler<undefined>;
   points: Point[];
   innerRings?: (Point[])[];
